Moves TestLifecycleAwareListener in a separate file
